Finally fixed the logout issue by making all pages ServerComponents and interactive components as ClientComponents.
